% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBP700_ERROR_ESTIMATION.R
\name{BBP700_ERROR_ESTIMATION}
\alias{BBP700_ERROR_ESTIMATION}
\title{BBP700_ERROR_ESTIMATION}
\usage{
BBP700_ERROR_ESTIMATION(filenc, BBP700_ADJUSTED, ERROR, scale_BBP700)
}
\arguments{
\item{filenc}{An object of class ncdf4 of argo bfile}

\item{BBP700_ADJUSTED}{vector of adjusted bbp700 values (numeric)}

\item{ERROR}{vector of bbp700 error values (numeric)}

\item{scale_BBP700}{factory calibration (scale factor) (numeric)}
}
\value{
vector of adjusted bbp700 error values (numeric)
}
\description{
This function estimates BPP700 error according to
https://archimer.ifremer.fr/doc/00491/60262/
}
