% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHLA_ERROR_ESTIMATION.R
\name{CHLA_ERROR_ESTIMATION}
\alias{CHLA_ERROR_ESTIMATION}
\title{CHLA_ERROR_ESTIMATION}
\usage{
CHLA_ERROR_ESTIMATION(filenc, CHLA_ADJUSTED, ERROR)
}
\arguments{
\item{filenc}{An object of class ncdf4 of argo bfile}

\item{CHLA_ADJUSTED}{vector of chlorophyll a adjusted values (numeric)}

\item{ERROR}{chlorophyll a error values (numeric)}
}
\value{
vector of chlorophyll a error values (numeric)
}
\description{
This function estimates chlorophyll a error according to
https://doi.org/10.13155/35385
}
