% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOXY_to_PPOX.R
\name{DOXY_to_PPOX}
\alias{DOXY_to_PPOX}
\title{DOXY_to_PPOX}
\usage{
DOXY_to_PPOX(PRES_DOXY, TEMP_DOXY, PSAL_DOXY, DOXY_DOXY)
}
\arguments{
\item{PRES_DOXY}{vector of pressure measured at the same
time as oxygen (numeric)}

\item{TEMP_DOXY}{vector of temperature values at the oxygen pressure level}

\item{PSAL_DOXY}{vector of salinity values at the oxygen pressure level}

\item{DOXY_DOXY}{vector of oxygen values in µmol/kg
at the oxygen pressure level}
}
\value{
vector of oxygen values in dbar (numeric)
}
\description{
This function convert oxygen concentration in µmol/kg into
partial pressure oxygen in dbar according to Bittig et al.
SCOR WG 142 https://doi.org/10.13155/45915
}
