\encoding{utf-8} 
\name{g1}
\alias{g1}
\docType{data}
\title{generic name for network X}
\description{This is the generic name for the original network X currently in use by the functions of this package.}
%\usage{data(sample1)}
%\format{
%  A data frame with 14 observations on the following 2 variables.
%  \describe{
%    \item{g1}{An original igraph network}
%    \item{g2}{An augmented igraph network}
%  }
%}

