\encoding{ISO-8859-2} 
\name{g2}
\alias{g2}
\docType{data}
\title{generic name for network Y}
\description{This is the generic name for the augmented network Y currently in use by the functions of this package.}
%\usage{data(sample1)}
%\format{
%  A data frame with 14 observations on the following 2 variables.
%  \describe{
%    \item{g1}{An original igraph network}
%    \item{g2}{An augmented igraph network}
%  }
%}

