% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPAR_15_DEPTH.R
\name{IPAR_15_DEPTH}
\alias{IPAR_15_DEPTH}
\title{IPAR_15_DEPTH}
\usage{
IPAR_15_DEPTH(PRES_CTD, PAR)
}
\arguments{
\item{PRES_CTD}{vector of pressure values at temperature pressure level}

\item{PAR}{vector of PAR}
}
\value{
a value of pressure
}
\description{
This function estimates the depth where PAR is <= 15
according to https://doi.org/10.1364/OE.26.024734
}
