% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLD.R
\name{MLD}
\alias{MLD}
\title{MLD}
\usage{
MLD(PRES_CTD, PSAL_CTD, TEMP_CTD)
}
\arguments{
\item{PRES_CTD}{vector of pressure values at temperature pressure level}

\item{PSAL_CTD}{vector of salinity values at temperature pressure level}

\item{TEMP_CTD}{vector of temperature values at temperature pressure level}
}
\value{
a value of mixed layer depth in meter
}
\description{
This function estimates the mixed layer depth according to
Boyer de Montegut et al. JGR, 2004 https://doi.org/10.1029/2004JC002378
}
