% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PH_adj.R
\name{PH_adj}
\alias{PH_adj}
\title{PH_adj}
\usage{
PH_adj(
  filenc_core,
  filenc,
  PARAM_name,
  OFFSET,
  SLOPE,
  DRIFT,
  ERROR,
  profile_date_juld,
  launch_date_juld,
  index_param
)
}
\arguments{
\item{filenc_core}{An object of class ncdf4 of argo core file}

\item{filenc}{An object of class ncdf4 of argo bfile}

\item{PARAM_name}{name of the variable to be adjusted (string)}

\item{OFFSET}{offset value for the oxygen adjustment (numeric)}

\item{SLOPE}{slope value for the oxygen adjustment (numeric)}

\item{DRIFT}{drift value for the oxygen adjustment (numeric)}

\item{ERROR}{vector pH error values (numeric)}
}
\value{
list with vector pH adjusted values (numeric)
and CTD flag availability (TRUE or FALSE)
}
\description{
This function estimates the pH adjustement according
to https://archimer.ifremer.fr/doc/00866/97828/
}
