% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PH_ERROR_ESTIMATION.R
\name{PH_ERROR_ESTIMATION}
\alias{PH_ERROR_ESTIMATION}
\title{PH_ERROR_ESTIMATION}
\usage{
PH_ERROR_ESTIMATION(
  filenc,
  PH_IN_SITU_TOTAL,
  PH_IN_SITU_TOTAL_ADJUSTED,
  ERROR,
  index_param,
  FLAG_CTD
)
}
\arguments{
\item{filenc}{An object of class ncdf4 of argo bfile}

\item{PH_IN_SITU_TOTAL}{vector of ph total raw values (string)}

\item{PH_IN_SITU_TOTAL_ADJUSTED}{vector of ph total adjusted values (string)}

\item{ERROR}{vector pH error values (numeric)}

\item{index_param}{parameter index in the file matrix}

\item{FLAG_CTD}{TRUE or FALSE for CTD availability}
}
\value{
vector of pH adjusted error values (numeric)
}
\description{
This function estimates the pH error according
to https://archimer.ifremer.fr/doc/00866/97828/
}
