% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPOX_to_DOXY.R
\name{PPOX_to_DOXY}
\alias{PPOX_to_DOXY}
\title{PPOX_to_DOXY}
\usage{
PPOX_to_DOXY(PRES_DOXY, TEMP_DOXY, PSAL_DOXY, PPOX_DOXY)
}
\arguments{
\item{PRES_DOXY}{vector of pressure measured at the same
time as oxygen (numeric)}

\item{TEMP_DOXY}{vector of temperature values at the oxygen pressure level}

\item{PSAL_DOXY}{vector of salinity values at the oxygen pressure level}

\item{PPOX_DOXY}{vector of oxygen values in dbar at the oxygen pressure level}
}
\value{
vector of oxygen values in µmol/kg (numeric)
}
\description{
This function convert partial pressure oxygen in dbar into
oxygen concentration in µmol/kg according to Bittig et al.
SCOR WG 142 https://doi.org/10.13155/45915
}
