% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/READ_CTD.R
\name{read_CTD}
\alias{read_CTD}
\title{read_CTD}
\usage{
read_CTD(filenc_core)
}
\arguments{
\item{filenc_core}{An object of class ncdf4 of argo core file}
}
\value{
a list of 3 vector: 1. pressure, 2. salinity, 3.temperature
}
\description{
This function read CTD argo file
}
