% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningFilter.R
\name{RunningFilter}
\alias{RunningFilter}
\title{RunningFilter}
\usage{
RunningFilter(k, y, na.fill = T, ends.fill = T, Method = "Average")
}
\arguments{
\item{k}{represents the half window, the window is set as (2k+1)}

\item{y}{is the list of points to be processed}

\item{na.fill}{is the parameter to indicate whether to process
the "NA" points, "T" means that all the "NA" points are included
in the function as well as valid points, and filled as valid values
unless there is not any valid values in the window (all methods are
applied through the parameter "na.rm=T"); "N" means that all the "NA"
points are excluded from the function and remained as "NA"
(all methods are processed only for valid points).}

\item{ends.fill}{is the parameter to indicate whether to process the
"marginal" points (the beginning and end points are not able to be
processed in a full window). "T" means these points are also processed,
but in a shorter window; "F" means these points are re-assigned as "NA".}

\item{Method}{is the parameter to indicate which method to be applied
in the list of points, "Average" means choosing the average value
in the window; "Median" means choosing the median value in the window;
"Maximum" and "Minimum" means choosing the maximal or minimal value
in the window.}
}
\value{
y value
}
\description{
RunningFilter is a function of running(moving/streaming)
filter, there are four types of method, Average and Median
methods are the filters to remove the high-frequency noise
through regulating the window range; Maximum and Minimum
methods are the functions to derive the nevelop lines
}
