\encoding{ISO-8859-2}
\name{SDDE.testing}
\alias{SDDE.testing}
\title{Testing}
\description{This function perform some tests to ensure validity of some SDDE functions. }
\usage{SDDE.testing()}

%\author{Etienne Lord, Margaux Le Cam, Eric Bapteste, Vladimir Makarenkov and Francois-Joseph Lapointe}


%\keyword{Augmented network, plot function}