\encoding{ISO-8859-2}
\name{SDDE.version}
\alias{SDDE.version}
\title{Version information.}
\description{This function return the current SDDE version. }
\usage{SDDE.version()}

%\author{Etienne Lord, Margaux Le Cam, Eric Bapteste, Vladimir Makarenkov and Francois-Joseph Lapointe}


%\keyword{Augmented network, plot function}