% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRITE_DM_BP.R
\name{WRITE_DM_BP}
\alias{WRITE_DM_BP}
\title{WRITE_DM_BP}
\usage{
WRITE_DM_BP(input_file)
}
\arguments{
\item{input_file}{text file including all usefull information
for the adjustment separated by ";"
filename             full path of the B file
filename_core        full path of the core file
metadata_filename    full path of the meta file
param                name of the variable to be adjusted
type                 A (ascending) D (descending) profiles
offset               offset coefficient
slope                slope coefficient
drift                drift coefficient
incline              incline T coefficient (for oxyen)
N_CYCLE_BEGIN        first Cycle Number for applying the drift
param_error          error values (in dbar for oxygen)
qc                   qc values for adjusted param
scientific_comment   string
date_update          date update yyyymmddhhmmss}
}
\value{
Bfile with param_adjusted, param_adjusted_qc, param_adjusted_error
}
\description{
This program is designed to modify a BR file into BD file
}
